requirements.txt
gymnasium>=0.29.1
numpy>=1.24
pyyaml>=6.0
matplotlib>=3.7
pandas>=2.0
tqdm>=4.66
stable-baselines3>=2.3.0
shimmy>=1.3.0
Pillow>=10.0
# Optional (disabled by default in config)
wandb>=0.16
